Update List.of.GRUB_INIT_TUNE.music.txt
We update to explicitly make it clear.

add GNU GPLv3 to make it clear we are Free as in Freedom. because this is coding right? but still, we still have this free as in beer as well.
acknowledge Chinese Bluetooth Audio Device firmware. A lot of Bluetooth speaker and headset may have the instance of the firmware.
mark the synthaxes
LeftRight while forgot the musical frequency
Scientific A because there are a conspiracy about 432 Hz where it is claimed as natural frequency. I am still researching it to find its benefit whether it was an over exageration hoacc or it is actually a real deal.